+++ client/src/game/ui/selection/shapecontext.vue
@@ -5,7 +5,9 @@ import Component from "vue-class-component";
 import { mapState } from "vuex";
 
 import ContextMenu from "@/core/components/contextmenu.vue";
+import Game from "@/game/Game.vue";
 import Prompt from "@/core/components/modals/prompt.vue";
+import SelectionBox from "@/core/components/modals/SelectionBox.vue";
 
 import { EventBus } from "@/game/event-bus";
 import { layerManager } from "@/game/layers/manager";
@@ -14,15 +16,16 @@ import { deleteShapes } from "../../shapes/utils";
 import { initiativeStore, inInitiative } from "../initiative/store";
 import { Layer } from "../../layers/layer";
 import { gameSettingsStore } from "../../settings";
-import Game from "@/game/Game.vue";
 import { Shape } from "@/game/shapes/shape";
 import { floorStore } from "../../layers/store";
 import { Floor } from "@/game/layers/floor";
 import { moveFloor, moveLayer } from "../../layers/utils";
+import { requestAssetOptions, sendAssetOptions } from "@/game/api/emits/asset";
 import { requestSpawnInfo, sendLocationChange } from "@/game/api/emits/location";
 import { sendShapesMove } from "@/game/api/emits/shape/core";
-import SelectionBox from "@/core/components/modals/SelectionBox.vue";
 import { ServerAsset } from "@/game/comm/types/shapes";
+import { AssetOptions } from "@/game/comm/types/asset";
+import { toTemplate } from "@/game/shapes/template";
 
 @Component({
     components: {
@@ -60,6 +63,7 @@ export default class ShapeContext extends Vue {
         this.$nextTick(() => (this.$children[0].$el as HTMLElement).focus());
     }
     close(): void {
+        if (this.$refs.prompt.visible || this.$refs.selectionbox.visible) return;
         this.visible = false;
     }
     getMarker(): string | undefined {
@@ -206,6 +210,44 @@ export default class ShapeContext extends Vue {
         if (gameStore.IS_DM) return true;
         return this.getSelection().every(s => s.ownedBy({ editAccess: true }));
     }
+    showDmNonSpawnItem(): boolean {
+        if (this.hasSpawnToken()) return false;
+        return gameStore.IS_DM;
+    }
+
+    hasAsset(): boolean {
+        return this.getActiveLayer()!
+            .getSelection()
+            .every(s => s.assetId !== undefined);
+    }
+
+    async saveTemplate(): Promise<void> {
+        const shape = this.getSelection()[0];
+        let assetOptions: AssetOptions = {
+            version: "0",
+            shape: shape.type,
+            templates: { default: {} },
+        };
+        if (shape.assetId) {
+            const response = await requestAssetOptions(shape.assetId);
+            if (response.success && response.options) assetOptions = response.options;
+        } else {
+            console.warn("Templates are currently only supported for shapes with existing asset relations.");
+            return;
+        }
+        const choices = Object.keys(assetOptions.templates);
+        try {
+            const choice = await this.$refs.selectionbox.open(this.$t("game.ui.templates.save").toString(), choices, {
+                defaultButton: this.$t("game.ui.templates.overwrite").toString(),
+                customButton: this.$t("game.ui.templates.create_new").toString(),
+            });
+            assetOptions.templates[choice] = toTemplate(shape.asDict());
+            sendAssetOptions(shape.assetId, assetOptions);
+        } catch {
+            // no-op ; action cancelled
+        }
+    }
+
     getLayerWord(layer: string): string {
         switch (layer) {
             case "map":
@@ -280,7 +322,6 @@ export default class ShapeContext extends Vue {
         <li @click="moveToFront" v-t="'game.ui.selection.shapecontext.move_front'"></li>
         <li @click="addInitiative" v-if="showInitiative()">{{ getInitiativeWord() }}</li>
         <li @click="deleteSelection" v-if="showDelete()" v-t="'game.ui.selection.shapecontext.delete_shapes'"></li>
-        <li v-if="hasSingleShape()" @click="openEditDialog" v-t="'game.ui.selection.shapecontext.show_props'"></li>
         <template v-if="hasSingleShape()">
             <li
                 v-if="markers.includes(getMarker())"
@@ -289,6 +330,12 @@ export default class ShapeContext extends Vue {
             ></li>
             <li v-else @click="setMarker" v-t="'game.ui.selection.shapecontext.set_marker'"></li>
         </template>
+        <li
+            @click="saveTemplate"
+            v-if="hasSingleShape() && showDmNonSpawnItem() && hasAsset()"
+            v-t="'game.ui.templates.save'"
+        ></li>
+        <li v-if="hasSingleShape()" @click="openEditDialog" v-t="'game.ui.selection.shapecontext.show_props'"></li>
     </ContextMenu>
 </template>