+++ client/src/game/ui/tools/createtoken_modal.vue
@@ -10,7 +10,7 @@ import Modal from "@/core/components/modals/modal.vue";
 
 import { calcFontScale } from "@/core/utils";
 import { LocalPoint } from "@/game/geom";
-import { CircularToken } from "@/game/shapes/circulartoken";
+import { CircularToken } from "@/game/shapes/variants/circulartoken";
 import { gameStore } from "@/game/store";
 import { getUnitDistance, l2g } from "@/game/units";
 import { Watch } from "vue-property-decorator";
@@ -65,8 +65,7 @@ export default class CreateTokenModal extends Vue {
             getUnitDistance(gameSettingsStore.unitSize / 2),
             this.text,
             "10px serif",
-            this.fillColour,
-            this.borderColour,
+            { fillColour: this.fillColour, strokeColour: this.borderColour },
         );
         token.addOwner({ user: gameStore.username, access: { edit: true } }, false);
         layer.addShape(token, SyncMode.FULL_SYNC, InvalidationMode.WITH_LIGHT);